requirements.txt
cairocffi==1.0.2
CairoSVG==2.3.0
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.7
django-braces==1.11.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-easy-pdf==0.1.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-paypal==0.5.0
et-xmlfile==1.0.1
gunicorn==19.7.1
html5lib==1.0.1
idna==2.8
jdcal==1.3
odfpy==1.3.5
openpyxl==2.4.8
Pillow==5.4.1
pycparser==2.19
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2018.9
PyYAML==3.13
reportlab==3.5.13
requests==2.21.0
six==1.12.0
sqlparse==0.2.4
tablib==0.12.1
tinycss2==0.6.1
unicodecsv==0.14.1
urllib3==1.24.1
WeasyPrint==47
webencodings==0.5.1
whitenoise==3.3.1
xhtml2pdf==0.2.3
xlrd==1.1.0
xlwt==1.3.0